de-tailwind and move files
