+++ client/scripts/views/sublayout.tsx
@@ -8,7 +8,6 @@ import app from 'state';
 import { handleEmailInvites } from 'views/components/header/invites_menu';
 import { Sidebar } from 'views/components/sidebar';
 import { MobileHeader } from 'views/mobile/mobile_header';
-import { LandingPageFooter } from 'views/pages/landing/landing_page_footer';
 import { SearchBar } from './components/search_bar';
 import { SublayoutHeaderLeft } from './sublayout_header_left';
 import { SublayoutHeaderRight } from './sublayout_header_right';
@@ -16,6 +15,7 @@ import { isNotUndefined } from '../helpers/typeGuards';
 import { TokenHero } from './token_hero';
 import { TokenTerms } from './token_terms';
 import { SidebarQuickSwitcher } from './components/sidebar/sidebar_quick_switcher';
+import { Footer } from './footer';
 
 type SublayoutAttrs = {
   alwaysShowTitle?: boolean; // show page title even if app.chain and app.community are unavailable
@@ -109,9 +109,7 @@ class Sublayout implements m.ClassComponent<SublayoutAttrs> {
                 {vnode.children}
                 {isNotUndefined(rightContent) && rightContent}
               </div>
-              {!app.isCustomDomain() && (
-                <LandingPageFooter list={footercontents} />
-              )}
+              {!app.isCustomDomain() && <Footer list={footercontents} />}
             </div>
           </div>
         </div>
