--- client/styles/pages/landing/landing_page_footer.scss
@@ -1,9 +1,0 @@
-@import 'client/styles/shared';
-
-.FooterNavsLinks.mb-2 {
-  a.text-gray-500 {
-    color: $text-color-black;
-    text-decoration: none;
-    cursor: pointer;
-  }
-}
